requirements.txt
fastapi~=0.115.6
PyYAML~=6.0.2
pydantic~=2.10.5
requests~=2.32.3
pyyaml~=6.0.2
uvicore